Move map-related tasks into mapping_tasks
